usage of long instead of int

+++ AtgScriptsExtension/ScreenshotToClipboardScript.cs
@@ -217,7 +217,7 @@ namespace AtgScriptsExtension
             {
                 case "bgr0":
                     bmp = new Bitmap((int)w, (int)h, PixelFormat.Format32bppRgb);
-                    bmp.Read32RgbFromPaddedRgb0((int)stride, ba.data, (int)ba.size.ToUInt64());
+                    bmp.Read32RgbFromPaddedRgb0(stride, ba.data, (long)ba.size.ToUInt64());
                     break;
                 default:
                     throw new ArgumentException($"Unsupported color format: {format}");
